Use zpty and zle -F instead of signals and &!
This fires off each async runner with its own pty, and uses zpty to manage
them. zle -F is used to evaluate the results from each runner and redraw
the prompt; this has the advantage that it will run only when the shell is
otherwise idle.

This means we no longer need to protect prompt drawing with a mutex, nor do
we need temporary files to communicate with the async processes.